STORE-367 Auto Sync User Profiles with External LDAP refractor to move sync to the correct service
